requirements.txt
fastapi
openai
uvicorn
python-dotenv
langchain
langchain_community
langchain-chroma  
langchain-huggingface
langchain-core
langchain-openai
qdrant-client
python-multipart
pypdf
streamlit
requests
